+++ common/documentLoader.js
@@ -1,6 +1,6 @@
 import * as DidJwk from '@digitalbazaar/did-method-jwk';
 import * as DidKey from '@digitalbazaar/did-method-key';
-import fetch from 'node-fetch';
+import * as DidWeb from '@digitalbazaar/did-method-web';
 
 import {
   CONTEXT as CRED_CONTEXT,
@@ -31,22 +31,12 @@ import {
   CONTEXT_URL as VDL_BASE_CONTEXT_URL
 } from '@digitalbazaar/vdl-context';
 import {CachedResolver} from '@digitalbazaar/did-io';
-import {CryptoLD} from 'crypto-ld';
-import {getResolver as didWebResolver} from 'web-did-resolver';
 import {Ed25519VerificationKey2020}
   from '@digitalbazaar/ed25519-verification-key-2020';
 import {JsonLdDocumentLoader} from 'jsonld-document-loader';
-import {parse} from 'did-resolver';
 import X25519KeyAgreement2020Context from 'x25519-key-agreement-2020-context';
-import {X25519KeyAgreementKey2020}
-  from '@digitalbazaar/x25519-key-agreement-key-2020';
 
-const cryptoLd = new CryptoLD();
-cryptoLd.use(Ed25519VerificationKey2020);
-cryptoLd.use(X25519KeyAgreementKey2020);
-const didWebDriver = options => {
-  return didWebResolver(options).web;
-};
+const didWebDriver = DidWeb.driver();
 const didKeyDriver = DidKey.driver();
 didKeyDriver.use({
   name: 'Ed25519',
@@ -61,12 +51,9 @@ didJwkDriver.use({
 });
 
 export const didResolver = new CachedResolver();
-didResolver.use({method: 'web', get: async ({did}) => {
-  const parsedDID = parse(did);
-  return didWebDriver()(did, parsedDID);
-}});
 didResolver.use(didKeyDriver);
 didResolver.use(didJwkDriver);
+didResolver.use(didWebDriver);
 
 const getDocumentLoader = () => {
   const jsonLdDocLoader = new JsonLdDocumentLoader();